vocode/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
azure-cognitiveservices-speech==1.38.0
beautifulsoup4==4.12.3
bs4==0.0.2
cartesia==1.0.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
elevenlabs==1.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
iterators==0.2.0
janus==1.0.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
miniaudio==1.61
multidict==6.0.5
nltk==3.8.1
openai==1.39.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==24.1
phonenumbers==8.13.42
pluggy==1.5.0
protobuf==4.25.4
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyht==0.0.28
PyPDF2==3.0.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
sentry-sdk==2.12.0
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
starlette==0.37.2
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
vocode==0.1.114a1
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
